# utilitatsfirma-api-interna/openapi/openapi_comanda_v1.json
{
  "openapi" : "3.0.1",
  "servers" : [ {
    "url" : "/utilitatsfirmaapi/interna",
    "description" : "Servidor local (sense protocol ni host) per a desplegaments en entorns on el context de l'aplicació és /utilitatsfirmaapi/interna)"
  }, {
    "url" : "http://localhost:8080/utilitatsfirmaapi/interna"
  }, {
    "url" : "https://dev.caib.es/utilitatsfirmaapi/interna"
  }, {
    "url" : "https://proves.caib.es/utilitatsfirmaapi/interna"
  }, {
    "url" : "https://se.caib.es/utilitatsfirmaapi/interna"
  }, {
    "url" : "https://www.caib.es/utilitatsfirmaapi/interna"
  } ],
  "paths" : {
    "/secure/estadistiques/v1" : {
      "get" : {
        "operationId" : "estadistiques",
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RegistresEstadistics"
                }
              }
            }
          }
        },
        "security" : [ {
          "BasicAuth" : [ ]
        } ]
      }
    },
    "/secure/estadistiques/v1/info" : {
      "get" : {
        "operationId" : "estadistiquesInfo",
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EstadistiquesInfo"
                }
              }
            }
          }
        },
        "security" : [ {
          "BasicAuth" : [ ]
        } ]
      }
    },
    "/secure/estadistiques/v1/of/{data}" : {
      "get" : {
        "operationId" : "estadistiquesPerData",
        "parameters" : [ {
          "name" : "data",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RegistresEstadistics"
                }
              }
            }
          }
        },
        "security" : [ {
          "BasicAuth" : [ ]
        } ]
      }
    },
    "/secure/estadistiques/v1/from/{dataInici}/to/{dataFi}" : {
      "get" : {
        "operationId" : "estadistiquesPerRang",
        "parameters" : [ {
          "name" : "dataInici",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "dataFi",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/RegistresEstadistics"
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "BasicAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Dimensio" : {
        "type" : "object",
        "properties" : {
          "codi" : {
            "type" : "string"
          },
          "valor" : {
            "type" : "string"
          }
        }
      },
      "Fet" : {
        "type" : "object",
        "properties" : {
          "codi" : {
            "type" : "string"
          },
          "valor" : {
            "type" : "number",
            "format" : "double"
          }
        }
      },
      "RegistreEstadistic" : {
        "type" : "object",
        "properties" : {
          "dimensions" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Dimensio"
            }
          },
          "fets" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Fet"
            }
          }
        }
      },
      "RegistresEstadistics" : {
        "required" : [ "temps" ],
        "type" : "object",
        "properties" : {
          "temps" : {
            "type" : "string",
            "format" : "date-time"
          },
          "fets" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RegistreEstadistic"
            }
          }
        }
      },
      "DimensioDesc" : {
        "required" : [ "codi", "nom" ],
        "type" : "object",
        "properties" : {
          "codi" : {
            "maxLength" : 32,
            "minLength" : 1,
            "type" : "string"
          },
          "nom" : {
            "maxLength" : 64,
            "minLength" : 1,
            "type" : "string"
          },
          "descripcio" : {
            "type" : "string"
          },
          "valors" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "EstadistiquesInfo" : {
        "required" : [ "codi", "dimensions", "indicadors" ],
        "type" : "object",
        "properties" : {
          "codi" : {
            "maxLength" : 2147483647,
            "minLength" : 1,
            "type" : "string"
          },
          "dimensions" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/DimensioDesc"
            }
          },
          "indicadors" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/IndicadorDesc"
            }
          },
          "versio" : {
            "type" : "string"
          },
          "data" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "IndicadorDesc" : {
        "required" : [ "codi", "nom" ],
        "type" : "object",
        "properties" : {
          "codi" : {
            "maxLength" : 32,
            "minLength" : 1,
            "type" : "string"
          },
          "nom" : {
            "maxLength" : 64,
            "minLength" : 1,
            "type" : "string"
          },
          "descripcio" : {
            "type" : "string"
          },
          "format" : {
            "type" : "string",
            "enum" : [ "LONG", "DECIMAL", "PERCENTAGE", "CURRENCY", "DATE", "DATETIME", "BOOLEAN" ]
          }
        }
      },
...
    },
    "securitySchemes" : {
      "BasicAuth" : {
        "type" : "http",
        "scheme" : "basic"
      }
    }
  }
}